backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tensorflow==2.15.0
numpy==1.26.2
pillow==10.1.0
opencv-python==4.8.1.78
albumentations==1.3.1
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
chromadb==0.4.18
httpx==0.25.2
python-json-logger==2.0.7
slowapi==0.1.9
